refactor executor logic out of server
